fix: remove home page change to login page.
