+++ webapp/src/components/Header.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { useTranslation } from "@/app/i18n/client";
+import { useCurrentUser } from "@/hooks/use-current-user";
 import { useSession } from "next-auth/react";
 import { usePathname, useRouter } from "next/navigation";
 import React, { useState } from "react";
@@ -9,11 +10,10 @@ import toast from "react-hot-toast";
 export default function Header({ lng }: { lng: string }) {
   const { t } = useTranslation(lng, "make-appointment");
   const router = useRouter();
-
-  const { data: session } = useSession();
+  const currentUser = useCurrentUser();
+  
 
   const currentPath = usePathname().substring(3);
-  const [loggedIn, setLoggedIn] = useState(false);
 
   const handleSwitchLanguage = () => {
     if (lng === "en") {
@@ -36,12 +36,12 @@ export default function Header({ lng }: { lng: string }) {
 
       {/* nav links */}
       <div className="hidden md:flex flex-grow px-8 py-2 justify-end gap-8">
-        <button
+        {/* <button
           onClick={() => router.push(`/${lng}`)}
           className="px-2 rounded-md font-semibold hover:bg-gray-300 hover:text-black transition-all"
         >
           {t("home")}
-        </button>
+        </button> */}
         <button
           onClick={() => router.push(`/${lng}/activity`)}
           className="px-2 rounded-md font-semibold hover:bg-gray-300 hover:text-black transition-all"
@@ -59,25 +59,25 @@ export default function Header({ lng }: { lng: string }) {
       {/* login */}
       {/* add w-1/12 if want fixed */}
       <div className="hidden md:flex justify-center">
-        {session != null ? (
+        {currentUser != null ? (
           <div className="flex p-2 gap-2 items-center rounded-md bg-neutral-200 shadow-md">
             <img
-              src={session.user?.image!}
+              src={currentUser.image!}
               alt="avatar"
               className="size-8 rounded-full shadow-md"
             />
-            <p className="text-black font-semibold">{session.user?.name}</p>
+            <p className="text-black font-semibold">{currentUser.name}</p>
           </div>
         ) : (
           <div className="flex gap-4 py-2">
             <button
-              onClick={() => router.push(`/${lng}/auth`)}
+              onClick={() => router.push(`/${lng}`)}
               className="px-2 rounded-md font-semibold bg-black hover:bg-gray-300 hover:text-black transition-all"
             >
               {t("login")}
             </button>
             <button
-              onClick={() => router.push(`/${lng}/auth`)}
+              onClick={() => router.push(`/${lng}`)}
               className="px-2 rounded-md font-semibold bg-white hover:bg-gray-300 text-black transition-all"
             >
               {t("signup")}
