+++ webapp/src/hooks/use-current-user.ts
@@ -1,0 +1,7 @@
+import { useSession } from "next-auth/react";
+
+export const useCurrentUser = () => {
+  const session = useSession();
+
+  return session.data?.user;
+};
