+++ webapp/src/hooks/use-session.ts
@@ -1,0 +1,7 @@
+import { useSession } from "next-auth/react";
+
+export const useCurrentSession = () => {
+  const session = useSession();
+
+  return session;
+};
